Fix Bug 118
Not sure why it was doing X+1, but in refactoring it to move pairs as they were picked; the issue went away.  This refactor also prevented you from picking the same card twice when there are 3 copies in the trash.
Added a special string to Character Card for tricks in trash.