Credit direct healing from non-target cards to that deck's heroes
For now, this credits healing to all characters in a multi-character
deck, despite conceptually Medico being the source of their healing.

This also behaves weirdly if a non-target hero card becomes a target
from another effect. For example, Doc Havoc's Unstable Serum would
normally credit the Doc for healing. But if he had the Omni-Unity
mission reward in Oblivaeon mode, Unstable Serum becomes a target so
takes credit for the healing itself.