Use a Yes/No when H-1 targets with the lowest HP is ambiguous
Instead of asking the player to enumerate the H-1 targets each time a
possible member of that set is being damaged.

This makes the UI more consistent with how "the (single) target with
the lowest HP is immune to damage", though the available decision text
isn't exactly correct.

It also fixes #258 by specifying a reasonable choice (assume yes, like
the helpers for single highest/lowest) in preview mode.