Modified Tainted Blood Trigger Timing for Out of Order Phases
Previously, the trigger was tied to  the going to the End Phase, as that satisfies the criteria normally.

However if there is a card out that changes the turn order, such as Toppling Skyscraper in Superstorm Akela, this no longer was accurate. Changed to be tied to any phase change where the FromPhase was DrawCard