Option 1
This is one idea for how to resolve "Affects"

This allows for things to affect the damage like Ta Moko and MistForm.

However, it causes things like "villain stun bolt" and mass levitation and Twisting Passages to not work. Some of those can be resolved by making Catacombs an indestructible target, but not all of them