Fix for Superstorm Akela Flip Bug
When the environment needs to refresh the UI when a card moves positions, it does this by flipping the card and then flipping it back. It chooses a random card to be the CardSource for this flip. It was coded to never choose the card that was calling RefreshUI as the source, but I think the correct behavior should be to never choose the card that is being flipped as the CardSource. With the original behavior, it would occasionally choose the card that flipped, which would then be unable to flip back to the front side because it can't cause things while flipped. In unit tests, this would occasionally cause the test TestRideTheCurrents_MoveTargetsAround to fail.