Fix for TestPrehistoricBehemothEndDamage
This test has La Capitan as the villain, and when going to the environment turn, it would activate the end of turn card play on La Paradoja Magnifica, which would sometimes play crew cards that deal damage and change which targets are expected to have the highest HP for the Behemoth damage. This fix destroys La Paradoja Magnifica to prevent this.