Fix for TestTakIshmael
This test uses Hyrian Snipe as the specimen played by Tak Ishmael at the end of the environment turn, which causes him to deal 1 damage. However, the test also has Menagerie play Aquatic Sphere with her end of turn effect, which puts the top card of the villain deck face down underneath itself. If that card happens to be Hyrian Snipe, it messes up the rest of the test. This fix stacks the deck so that Lumobat Flock will end up underneath Aquatic Sphere.