Fix for TestSwarmingProtocolCyberintegration
This test uses Swarming Protocol Cypher's power to put a face-down card next to Ra, as well as Dermal Aug, and Retinal Aug next to Tachyon. If not specified, the test chooses a card at random from Cypher's hand to play face-down, and if this happens to be Dermal Aug or Retinal Aug, the test fails. This fix ensures that Initiated Upgrade will always be chosen as the face-down card.