Fix TestWetWork_Oblivaeon
This test starts by moving Haka to BZ2, then discards the top card of OblivAeon's deck. If that card happened to be The Abyss Stares Back, it would move everyone over to BZ2, which means that Haka would be visible to Tango One when he should not be for this test. Fixed by stacking OblivAeon's deck with Focus of Power before discarding.