Fix TestGho5t_Destroyed_OptionalDraw
This test uses Interpolation Beam to check which heroes drew a card from Gho5t getting destroyed. Since Omnitron plays a card at the end of his turn, there would sometimes be 2 Interpolation Beams in play, which meant that the damage was doubled. Fixed by putting Interpolation Beam on top of the deck instead of playing it.