Fix for TestWastelandRoninCricketIncap2_Oblivaeon_EdgeCase6
This test cycles through several rounds of an OblivAeon game, which activates all end-of-turn effects. If Aeon Master happens to be the BZ2 scion, he can cause a decision that breaks the test due to the game trying to select the Hairdryer when the only selectable cards are Aeon Men. Fixed by resetting decision logic before cycling through the turns a second time.