Fix for TestGrasshopperKick
This test uses Plummeting Monorail to test the immunity effect of Grasshopper Kick, but playing Monorail also plays the top card of Akash'Bhuta's deck. If she happens to play Disrupt the Field, Monorail is destroyed and messes up the test. Fixed by stacking Akash's deck with Allies of the Earth.