Fix TestMenagerieFrontChallenge
This test needs Prized Catch to go next to Scholar, and it tries to ensure that by discarding the other heroes' hands before starting the game. However, if Scholar happens to have no one-shots in his hand, the game will instead put it on the Sentinels. Fixed by putting Don't Dismiss Anything in the Scholar's hand before starting the game.