Fix for TestDynamo_Advanced
Dynamo's advanced rules discard the top card of his deck at the start of the game, and this test then plays Copperhead to discard another card. But if the first card discarded from the top of the deck was Copperhead, the card count in the trash is incorrect after Copperhead is played. Fixed by putting Crime Spree on top of the villain deck before starting the game.