Fix FlashReconOblivAeon
For this test, Flash Recon is supposed to reveal the top card of all three hero decks, all three villain decks, and the environment deck. To make sure Vanish can see the Scion deck, it plays Borr into BZ1. But if Borr happened to be the randomly picked Scion in BZ2, he can't be played into BZ1, so there is no Scion there. Fixed by removing the part where it plays Borr into BZ1, and instead fixing the shield identifier in SetupGameController so that the shield actually enters play, and Voidsoul is in BZ1 for the test.

I also stacked OblivAeon's deck with Focus of Power just in case revealing The Abyss Stares Back causes other issues.