Fix TestUnstablePower
This test requires a Cascade card in Pyre's trash to shuffle into his deck. If a cascade card happens to be in Pyre's hand at the start of the game, there will be 2 cascade cards in the trash instead. This means instead of the cascade card the test expects to get shuffled, the other cascade card gets shuffled, and the test fails. Fixed by moving all cascade cards to Pyre's deck after Pyre's hand is discarded.