Fix TestPillarsOfCreation_Timing_BreakingTheRules
This test has Wager Master as the villain, so random game-overs can cause it to fail. The test did account for this by putting Losing to the Odds in the trash, but this doesn't work if Losing to the Odds was one of the initial Condition cards played by Wager Master's setup. Fixed by putting it in the trash before starting the game so it can't be one of the initial conditions.

I also found that if all 3 initial conditions are Wagelings, this causes an immediate game over. Fixed by also banishing the Wagelings before the game starts.