Fix TestAceOfSaintsShuffleSame2Cards
This test starts by moving Ace of Saints to Baccarat's hand and putting the rest of his deck in the trash. However, it assumes that Ace of Saints was moved to hand from the deck, and just discards the top 35 cards of the deck. If Ace of Saints started in the hand, there would be 36 cards in the deck, and one would be left over after moving the rest to the trash. This causes the count of cards in Baccarat's deck to be off, and the test fails. Fixed by using MoveAllCards() to empty the deck instead of discarding 35 cards.