Fix TestFallDealLightningNoDamage
This test uses Sonic Mine to make Lady unable to deal damage, but if both Sonic Mines are part of the starting set of devices for Ambuscade, for some reason trying to put Sonic Mine into play removes its triggers, and it doesn't deal damage when destroyed. Fixed by putting all Sonic Mines in the trash before starting the game.