Fix TestWarpBridge_Issue1016_NonRealCards
This test requires that a specific number of cards be available for the decision on Warp Bridge. Since Kaargra plays a card at the end of her turn, she can play a card that flips her, and she along with the gladiators in play can cause a hero to earn the title Unbreakable. When the field is cleared later to set the stage for the test, it only removes the titles in the title play area, not ones earned by targets, so the earned title is an extra card in the Warp Bridge decision and causes the test to fail. Fixed by stacking the deck with Perith so that Kaargra can never flip.