Fix TestLabyrinthGuideEnvironmentTurn_TwistingPassagesNotInPlay_NoAvailableHeroes
End-of-turn effects can happen and deal damage to heroes in between when their HP is stored and checked. Fixed by first going to the end of Scion 1's turn, then storing HP, then going to the start of the second environment turn. Also destroyed all Aeon Men just to be safe.