Fix TestAcceleratedEvolutionAnathemaEndOfTurnFront_ArmOnTop_Flip
When Anathema flips during this test, Explosive Transformation is shuffled from the trash into the deck. If it ends up on top of the deck and Reflex Booster was one of Anathema's starting cards, it will be played and flip Anathema back to the front side. Fixed by stacking the deck with Rampage after the shuffle.