Fix TestStartOfVillainOblivaeonTurn
Not sure why this test is even here, but it was flaky and I fixed it. If Voidsoul happened to be the BZ2 scion, he would play a card at the end of the turn, which had the chance to move OblivAeon to BZ2 and change which BZ Impending Doom gets played in. Fixed by stacking the Scion deck with Scatter Slaughter.