Fix flaky unit tests for Cypher, Starlight, Knight, and Baccarat
TestHeuristicAlgorithm_NoAugsInDeck - If Heuristic Algorithm happens to be played from Cypher's hand, the card count in his hand is off for the test. Fixed by playing Heuristic Algorithm from the trash instead.

Area51StarlightInnatePower - This test starts by discarding Starlight's hand. If one of the Pillars of Creation is discarded this way, then the copy of Pillars selected for the test and the copy selected when the power is used will be different copies, and the test will fail. Fixed by shuffling Starlight's trash into her deck after discarding her hand.

TestAceOfSinnersShuffleSame2Cards - Same fix as for TestAceOfSaintsShuffleSame2Cards in a previous pull request. It starts by putting Ace in hand and discarding Baccarat's deck, but it does this by discarding 35 cards. If Ace happened to be in hand at the start, there will be one card left in the deck. Fixed by using MoveAllCards() instead.

MaidensBlessing - This test starts by discarding The Knight's hand, and later looks for copies of Whetstone in the deck. If both copies of Whetstone happened to be in hand at the start, it will find no copies and the test will break. Fixed by shuffling Knight's trash into the deck after discarding his hand.