Fix unit test for Anathema
TestAcceleratedEvolutionAnathemaEndOfTurnFront_HeadOnTop_Flip - Same fix as for the test TestAcceleratedEvolutionAnathemaEndOfTurnFront_ArmOnTop_Flip in a previous test. After flipping, Explosive Transformation is shuffled into the villain deck, and if Reflex Booster was one of the starting cards for Anathema, it could get played at the end of the villain turn, flipping Anathema and failing the test. Fixed by stacking the deck with Rampage after the flip.