Fix for TestAcceleratedEvolutionHeadIndestructibilityTest_FlippedVillainTurn_NegativeHP
If Reflex Booster was one of Anathema's starting cards, it could play Explosive Transformation at the end of Anathema's turn. If it did, and a head was revealed, Reflex Booster would end up under Anathema's character card instead of in the trash where the test expects it to be. Fixed by stacking the villain deck with Rampage.