Fix for TestAcceleratedEvolutionHeadIndestructibilityTest_FlippedVillainTurn_NegativeHP
This test would fail if Reflex Booster was one of the starting cards and Explosive Transformation was played by Reflex Booster at the end of the villain turn and Explosive revealed a head. In that case, Reflex Booster would end up under Anathema's character card instead of in the trash, and the test would fail. Fixed by stacking the villain deck with Rampage