Fix unit tests for Tango One and Gyrosaur
TestHyperspinPlayDrawnCrashResponse - This test could fail if Wrecking Ball is in Gyrosaur's hand at the start, since it would get put on top of the deck and then drawn to activate Hyperspin, and the card count in hand would be wrong. Fixed by putting Wrecking Ball on top of the deck before storing the hand size.

TestCriticalHitDamageFail and TestCriticalHitDamageShinobiAssassin - These tests would store the second card from the top of Tango One's deck and check that it is on top of the deck after Critical Hit activates. If Critical Hit happened to be the second card from the top of the deck, it would get played for the test, and the test would fail. Fixed by stacking the deck so that Infiltrate is always the second card from the top.