Fix flaky tests for heroes
TestResourcefulDaydreamer - The test would mark the "top card" of Drift's hand for discard, which would sometimes be the same card Drift was supposed to play. Fixed by always having the discard be Past Focus.

MaidensBlessing - The test looks for Whetstone in The Knight's deck, but if all copies of Whetstone are in Knight's hand, it won't find any in the deck. Fixed by shuffling Knight's trash into his deck after his hand is discarded.

TestHyperspinPlayDrawnCrashResponse - The test would put Wrecking Ball on top of the deck after storing Gyrosaur's hand size. If Wrecking Ball was in her hand, the card count would be off. Fixed by stacking the deck before storing hand size.

TestHeuristicAlgorithm_NoAugsInDeck - The card count in Cypher's hand is off if Heuristic Algorithm is played from his hand. Fixed by moving it to the trash before storing hand size.

TestWastelandRoninCricketIncap2_Oblivaeon_EdgeCase2 - If Voidsoul is the BZ2 scion, he plays a card at the end of the Scion turn, which can mess up the test. Fixed by stacking the scion deck with Scatter Slaughter.

TestAceOfSinnersShuffleSame2Cards - Same fix as in a previous pull request. If Ace of Sinners started off in Baccarat's hand, the test would not completely empty his deck. Fixed by making sure it always empties his deck.

TestCriticalHitDamageFail - The test would get the second card from the top of Tango's deck and make sure it's on top after Critical Hit procs. If that card happened to be Critical Hit, it would be in play instead of on top, and the test would fail. Fixed by stacking the deck so Infiltrate is always the second card from the top of the deck.

TestCriticalHitDamageShinobiAssassin - same fix as TestCriticalHitDamageFail