Update FlashBeforeYourEyesCardController.cs
Fix Flash Before Your Eyes behavior in OblivAeon mode. Previously it would find the visible trashes based on the TurnTaker, but this meant that the scion and aeon men trashes were visible as long as Terminus was in the same BZ as OblivAeon. Fixed so it checks whether the trash itself is visible to Terminus.